Recents: Port "Development shortcut" (2/2)
This patch includes the equivalent of our Development Shortcuts in
CM11 recents. Recents is very different now and this is mostly a rewrite.

The semantics changed a bit as well. The current code already has special
developer behavior when long-pressing the app icon and that is also where
the developer shortcuts are hooked. If the app icon is long-pressed and
developer shortcuts are active, a menu will be shown with the several available
options.

Fixes CML-61

Change-Id: I391a360b3dd88e179c2315503d18125f8d0b3792

Conflicts:
	core/java/android/provider/Settings.java
	packages/SystemUI/res/values/cm_strings.xml

Conflicts:
	core/java/android/provider/Settings.java
	packages/SystemUI/res/values/deso_strings.xml

Conflicts:
	packages/SystemUI/res/values/deso_strings.xml

Conflicts:
	core/java/android/provider/Settings.java