Correct the lockscreen carrier text alignment and text color
In AOSP the text is left aligned, not centered, and the text color is white, not grey

Change-Id: Ie9c77761d3440336ae57abed24c8403bda06be3e

Correct the lockscreen carrier text case

The carrier text is no longer set to all caps in AOSP

Change-Id: I6a28a2d54a3b748f418af9f867b8be562911ef59

Clean up keyguard carrier text handling.

Return to AOSP code while still keeping the support for MSIM. Doing so
both simplifies the code and fixes the left alignment.

Change-Id: Icc48082818eccbaab80a9b96d7200e30458727dc