systemui: animate clear recents exit
Currently, the animation just clear all the tasks at the same time. Just add
a delay between every task remove call, and hide clear recents before start
any animation.

Change-Id: Idbdeacc68c6519d28dc875d09cd75c71e2e9f701
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>

systemui: never show excluded recent task in recents view

Actually, when there there isn't recent tasks to show, the systemui forces to show
task with the flag FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS, but only the first one.
This create an non-consistent behaviour when the user remove all recent task and
then there is a new recent task in the recents view, kill that and a new one
appears, an so on, until no task are available.

This patch changes that behaviour to:
- Never show tasks in recents view with the flag FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS
- Clear recents button kill all the tasks (included the ones with FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS)

Change-Id: I8dc2859ca9dcde7229d5fdb502da313b1ae0dccb
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/recents/views/TaskStackView.java

systemui: adjust position of clear recents button

If there is a statusbar then align to it. Otherwise aling to task views rather than top/right.
This patch fixed the position in tablets (align it with the search bar) and in landscape mode
(reducing the space between task views and the clear recents button), with is a more consistent
ui because the button is always properly aligned with the recent task views.

Change-Id: Ic867d5421cacba627c4b844db3d3715f20a4444c
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>

systemui: animate clear recents button on exit animation

Change-Id: Ic617f50400558be4ad92a4a5ea3ef441d2afde6a
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>

systemui: fix glitches with recents view and search widget

Reduce the glitches causes by switchin between the recents view and the search activity (with text
and voice search). Force search widget (specially to QuickSearchBox) to view the recents activity
as a HOME_SCREEN widget category instead of a SEARCHBOX widget category.

Change-Id: I08157d7de8dab806b19d2f4a28f9843402d12179
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>